% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcove.R
\name{rcove}
\alias{rcove}
\title{A Rcove Function}
\arguments{
\item{theta}{Angle about which to rotate current velocities. Degrees; Numeric}

\item{velocity}{Velocity time series; Vector consisting of two columns u,v.}

\item{release_pts}{Start locations for propagules; sf object}

\item{PD}{Planktonic propagule duration of species of interest; days}

\item{CPD}{Competent planktonic propagule duration of species of interest; days}

\item{land}{Land to be subtracted from final concatenated polygons; sf object}

\item{proj}{Projected crs in units m.}
}
\description{
This function allows you to Calculate Dispersal Range of Propagules using Rotated Components of Water Velocity
}
\examples{
rcove()
}
\keyword{dispersal}
